refactor(form-control): refactor FormControlMixin
