+++ packages/form-control/src/FormControlMixinLegacy.ts
@@ -1,14 +1,25 @@
-import { Constructor, CustomValidityState, FormControlInterface, FormValue, IControlHost, Validator } from './types';
+import { IElementInternals } from 'element-internals-polyfill';
+import { Constructor, FormControlInterface, FormValue, IControlHost, Validator } from './types';
+
+/** Traverse the prototype chain to check for a property descriptor */
+function getPropertyDescriptor(object: IControlHost, key: string): PropertyDescriptor|null {
+  if (object.constructor === HTMLElement) {
+    return null;
+  }
+  return Object.getOwnPropertyDescriptor(object, key) || getPropertyDescriptor(Object.getPrototypeOf(object), key)
+}
 
 export function FormControlMixin<
-  TBase extends Constructor<HTMLElement & IControlHost> & { observedAttributes?: string [] }
+  TBase extends Constructor<HTMLElement & IControlHost> & { observedAttributes?: string[] }
 >(SuperClass: TBase) {
   class FormControl extends SuperClass {
     /** Wires up control instances to be form associated */
-    static get formAssociated(): boolean {
+    static get formAssociated() {
       return true;
     }
-
+setValue(a: string): void {
+  console.log(a)
+}
     /**
      * A list of Validator objects that will be evaluated when a control's form
      * value is modified or optionally when a given attribute changes.
@@ -16,9 +27,9 @@ export function FormControlMixin<
      * When a Validator's callback returns false, the entire form control will
      * be set to an invalid state.
      */
-     declare static formControlValidators: Validator[];
+    declare static formControlValidators: Validator[];
 
-     /**
+    /**
      * If set to true the control described should be evaluated and validated
      * as part of a group. Like a radio, if any member of the group's validity
      * changes the the other members should update as well.
@@ -35,7 +46,7 @@ export function FormControlMixin<
      * validator will be evaluated whenever the host's required attribute
      * is updated.
      */
-     static get observedAttributes(): string[] {
+    static get observedAttributes(): string[] {
       const validatorAttributes = this.validators.map((validator) => validator.attribute);
 
       const observedAttributes = super.observedAttributes || [];
@@ -54,7 +65,7 @@ export function FormControlMixin<
     }
 
     /** The ElementInternals instance for the control. */
-    internals = this.attachInternals();
+    internals = this.attachInternals() as unknown as IElementInternals;
 
     /**
      * Keep track of if the control has focus
@@ -75,6 +86,11 @@ export function FormControlMixin<
      */
     #touched = false;
 
+    /** Will return true if the control has a checked property */
+    get #isCheckedElement(): boolean {
+      return 'checked' in this;
+    }
+
     /** All of the controls within a root with a matching local name and form name */
     get #formValidationGroup(): NodeListOf<FormControl> {
       const rootNode = this.getRootNode() as HTMLElement;
@@ -83,17 +99,272 @@ export function FormControlMixin<
     }
 
     /**
-     * Acts as a cache for the current value so the value can be re-evaluated
-     * whenever an attribute changes or on some other event.
+     * The element that will receive focus when the control's validity
+     * state is reported either by a form submission or via API
+     *
+     * We use declare since this is optional and we don't particularly
+     * care how the consuming component implements this (as a field, member
+     * or getter/setter)
      */
-    #value: FormValue = '';
+    declare validationTarget: HTMLElement | null;
+
+    /**
+     * The controls' form value. As this property is updated, the form value
+     * will be updated. If a given control has a `checked` property, the value
+     * will only be set if `checked` is truthy.
+     */
+    value: FormValue = '';
+
+    /** Return a reference to the control's form */
+    get form(): HTMLFormElement {
+      return this.internals.form;
+    }
+
+    /**
+     * Will return true if it is recommended that the control shows an internal
+     * error. If using this property, it is wise to listen for 'invalid' events
+     * on the element host and call preventDefault on the event. Doing this will
+     * prevent browsers from showing a validation popup.
+     */
+    get showError(): boolean {
+      return this.#shouldShowError();
+    }
+
+    /**
+     * Forward the internals checkValidity method
+     * will return the valid state of the control.
+     */
+    checkValidity(): boolean {
+      return this.internals.checkValidity();
+    }
+
+    /**
+     * The validation message shown by a given Validator object. If the control
+     * is in a valid state this should be falsy.
+     */
+    get validationMessage(): string {
+      return this.internals.validationMessage;
+    }
+
+    /**
+     * The element's validity state after evaluating the control's Validators.
+     * This property implements the same patterns as the built-in constraint
+     * validation strategy.
+     */
+    get validity(): ValidityState {
+      return this.internals.validity;
+    }
+
+    /* eslint-disable  @typescript-eslint/no-explicit-any */
+    constructor(...args: any[]) {
+      super(...args);
+      this.addEventListener('focus', this.#onFocus);
+      this.addEventListener('blur', this.#onBlur);
+      this.addEventListener('invalid', this.#onInvalid);
+    }
+
+    attributeChangedCallback(name: string, oldValue: string, newValue: string): void {
+      super.attributeChangedCallback?.(name, oldValue, newValue);
+
+      /**
+       * Check to see if a Validator is associated with the changed attribute.
+       * If one exists, call control's validate function which will perform
+       * control validation.
+       */
+      const proto = this.constructor as typeof FormControl;
+      const validator = proto.getValidator(name);
+
+      if (validator && this.validationTarget) {
+        /** Determine the value to validate */
+        let valueToValidate: FormValue = this.value;
+        if (this.#isCheckedElement && !this.checked) {
+          valueToValidate = null;
+        }
+        this.#validate(valueToValidate);
+      }
+    }
+
+    connectedCallback() {
+      super.connectedCallback?.();
+
+      /** Initialize the form control and perform initial validation */
+      this.#initFormControl();
+    }
+
+    disconnectedCallback() {
+      super.disconnectedCallback?.();
+
+      /**
+       * Remove the event listeners that toggles the touched and focused states
+       */
+      this.removeEventListener('focus', this.#onFocus);
+      this.removeEventListener('blur', this.#onBlur);
+      this.removeEventListener('invalid', this.#onInvalid);
+    }
+
+    /**
+     * Initialize the form control
+     * @private
+     */
+    #initFormControl(): void {
+      /** Closed over variable to track value changes */
+      let value: FormValue = this.value || '';
+
+      /** Look to see if '`checked'` is on the control's prototype */
+      const hasChecked = this.#isCheckedElement;
+
+      /**
+       * The FormControlMixin writes the value property on the element host
+       * this checks to see if some other object in the prototype chain
+       * has a getter/setter for value and saves a reference to those.
+       *
+       * We do this to make sure that we don't overwrite behavior of an object
+       * higher in the chain.
+       */
+      const descriptor = getPropertyDescriptor(this, 'value');
+
+      /** Make sure to defer to the parent */
+      const set = descriptor && descriptor.set;
+      const get = descriptor && descriptor.get;
+
+      /** Define the FormControl's value property */
+      Object.defineProperty(this.hasOwnProperty('value') ? this :  this.constructor.prototype, 'value', {
+        get() {
+          /** If a getter already exists, make sure to call it */
+          if (get) {
+            return get.call(this);
+          }
+          return value;
+        },
+        set(newValue) {
+          /** Save a reference to the new value to use later if necessary */
+          value = newValue;
+
+          this.#commitValue(newValue);
+
+          /** If a setter already exists, make sure to call it */
+          if (set) {
+            set.call(this, newValue);
+          }
+        }
+      });
+
+      /**
+       * If checked already exists on a prototype, we need to monitor
+       * for changes to that property to ensure the proper value is set on the
+       * control's form.
+       *
+       * TODO: Justin Fagnani cautioned that this might not scale well. Maybe
+       * this should be a direct check against the value of checked ...
+       */
+      if (hasChecked) {
+        /**
+         * As with value, save a reference to the getter/setter if they already
+         * exist in the prototype chain
+         */
+        const descriptor = getPropertyDescriptor(this, 'checked');
+        const get = descriptor && descriptor.get;
+        const set = descriptor && descriptor.set;
+
+        /** Close over the initial value to use in the new getter/setter */
+        let { checked } = this;
+
+        Object.defineProperty(this, 'checked', {
+          get() {
+            /** If a getter exists, use it */
+            if (get) {
+              return get.call(this);
+            }
+            return checked;
+          },
+          set(newChecked) {
+            /** Updated closure value */
+            checked = newChecked;
+
+            /** If a setter exists, use it */
+            if (set) {
+              set.call(this, newChecked);
+            }
+
+            this.#commitValue(this.value);
+          }
+        });
+      }
+
+      this.#commitValue(value);
+    }
+
+    /** Reset control state when the form is reset */
+    formResetCallback() {
+      this.resetFormControl();
+    }
+
+    /**
+     * A callback for when the controls' form value changes. The value
+     * passed to this function should not be confused with the control's
+     * value property, this is the value that will appear on the form.
+     * In cases where `checked` did not exist on the control's prototype
+     * upon initialization, this value and the value property will be identical;
+     * in cases where `checked` is present upon initialization, this will be
+     * effectively `this.checked && this.value`.
+     */
+    declare valueChangedCallback: (value: FormValue) => void;
+
+    /**
+     * Resets a form control to its initial state
+     */
+    resetFormControl(): void {
+      if (this.#isCheckedElement) {
+        this.checked = false;
+      } else {
+        this.value = '';
+      }
+      this.#touched = false;
+      this.#forceError = false;
+      this.#shouldShowError();
+    }
+
+    /**
+     * If the control has a checked property, make sure that it is
+     * truthy before setting the form control value. If it is falsy,
+     * remove the form control value.
+     */
+    #commitValue(value: FormValue): void {
+      if (this.#isCheckedElement && !this.checked) {
+        this.#setValue(null);
+      } else {
+        this.#setValue(value);
+      }
+    }
+
+    /**
+     * Check to see if an error should be shown. This method will also
+     * update the internals state object with the --show-error state
+     * if necessary.
+     * @private
+     */
+    #shouldShowError(): boolean {
+      if (this.hasAttribute('disabled')) {
+        return false;
+      }
+
+      const showError = this.#forceError || (this.#touched && !this.validity.valid && !this.#focused);
+
+      if (showError) {
+        this.internals.states.add('--show-error');
+      } else {
+        this.internals.states.delete('--show-error');
+      }
+
+      return showError;
+    }
 
     /**
      * Set this[touched] and this[focused]
      * to true when the element is focused
      * @private
      */
-     #onFocus = (): void => {
+    #onFocus = (): void => {
       this.#touched = true;
       this.#focused = true;
       this.#shouldShowError();
@@ -128,170 +399,36 @@ export function FormControlMixin<
       this?.validationMessageCallback?.(this.showError ? this.validationMessage : '');
     };
 
-    /** Return a reference to the control's form */
-    get form(): HTMLFormElement {
-      return this.internals.form;
-    }
-
     /**
-     * Will return true if it is recommended that the control shows an internal
-     * error. If using this property, it is wise to listen for 'invalid' events
-     * on the element host and call preventDefault on the event. Doing this will
-     * prevent browsers from showing a validation popup.
-     */
-    get showError(): boolean {
-      return this.#shouldShowError();
-    }
-
-    /**
-     * Forward the internals checkValidity method
-     * will return the valid state of the control.
-     */
-    checkValidity(): boolean {
-      return this.internals.checkValidity();
-    }
-
-    /** The element's validity state */
-    get validity(): ValidityState {
-      return this.internals.validity;
-    }
-
-    /**
-     * The validation message shown by a given Validator object. If the control
-     * is in a valid state this should be falsy.
-     */
-    get validationMessage(): string {
-      return this.internals.validationMessage;
-    }
-
-    /* eslint-disable  @typescript-eslint/no-explicit-any */
-    constructor(...args: any[]) {
-      super(...args);
-      this.addEventListener('focus', this.#onFocus);
-      this.addEventListener('blur', this.#onBlur);
-      this.addEventListener('invalid', this.#onInvalid);
-    }
-
-    attributeChangedCallback(name: string, oldValue: string, newValue: string): void {
-      super.attributeChangedCallback?.(name, oldValue, newValue);
-
-      /**
-       * Check to see if a Validator is associated with the changed attribute.
-       * If one exists, call control's validate function which will perform
-       * control validation.
-       */
-      const proto = this.constructor as typeof FormControl;
-      const validator = proto.getValidator(name);
-
-      if (validator && this.validationTarget) {
-        this.setValue(this.#value);
-      }
-    }
-
-    /** PUBLIC LIFECYCLE METHODS */
-
-    /**
-     * Sets the control's form value if the call to `shouldFormValueUpdate`
-     * returns `true`.
-     * @param value {FormValue} - The value to pass to the form
-     */
-    setValue(value: FormValue): void {
-      this.#value = value;
-      const valueShouldUpdate = this.shouldFormValueUpdate();
-      const valueToUpdate = valueShouldUpdate ? value : null;
-
-      this.internals.setFormValue(valueToUpdate as string);
-      this.#validate(valueToUpdate);
-    }
-
-    /**
-     * This method can be overridden to determine if the control's form value
-     * should be set on a call to `setValue`. An example of when a user might want
-     * to skip this step is when implementing checkbox-like behavior, first checking
-     * to see if `this.checked` is set to a truthy value. By default this returns
-     * `true`.
-     */
-    shouldFormValueUpdate(): boolean {
-      return true;
-    }
-
-    /** DECLARED INSTANCE METHODS AND PROPERTIES*/
-
-    /**
-     * This method is used to override the controls' validity message
-     * for a given Validator key. This has the highest level of priority when
-     * setting a validationMessage, so use this method wisely.
-     *
-     * The returned value will be used as the validationMessage for the given key.
-     * @param validationKey {string} - The key that has returned invalid
-     */
-    declare validityCallback: (validationKey: string) => string | void;
-
-     /**
-      * Called when the control's validationMessage should be changed
-      * @param message { string } - The new validation message
-      */
-    declare validationMessageCallback: (message: string) => void;
-
-    /**
-     * A callback for when the controls' form value changes. The value
-     * passed to this function should not be confused with the control's
-     * value property, this is the value that will appear on the form.
-     *
-     * In cases where `checked` did not exist on the control's prototype
-     * upon initialization, this value and the value property will be identical;
-     * in cases where `checked` is present upon initialization, this will be
-     * effectively `this.checked && this.value`.
-     */
-    declare valueChangedCallback: (value: FormValue) => void;
-
-    /**
-     * The element that will receive focus when the control's validity
-     * state is reported either by a form submission or via API
-     *
-     * We use declare since this is optional and we don't particularly
-     * care how the consuming component implements this (as a field, member
-     * or getter/setter)
-     */
-     declare validationTarget: HTMLElement | null;
-
-    /** PRIVATE LIFECYCLE METHODS */
-
-    /**
-     * Check to see if an error should be shown. This method will also
-     * update the internals state object with the --show-error state
-     * if necessary.
+     * Sets the control's value when updated and invokes the valueChangedCallback
+     * for the element. Once the value has been set, invoke the Validators.
      * @private
      */
-     #shouldShowError(): boolean {
-      if (this.hasAttribute('disabled')) {
-        return false;
+    #setValue(value: FormValue): void {
+      this.#forceError = false;
+      this.internals.setFormValue(value as string); /** Typed to string to work with polyfill */
+      if (this.valueChangedCallback) {
+        this.valueChangedCallback(value);
       }
-
-      const showError = this.#forceError || (this.#touched && !this.validity.valid && !this.#focused);
-
-      if (showError) {
-        this.internals.states.add('--show-error');
-      } else {
-        this.internals.states.delete('--show-error');
+      this.#validate(value);
+      const showError = this.#shouldShowError();
+      if (this.validationMessageCallback) {
+        this.validationMessageCallback(showError ? this.validationMessage : '');
       }
-
-      return showError;
     }
 
-     /**
+    /**
      * Call all the Validators on the control
      * @private
      */
     #validate(value: FormValue): void {
       const proto = this.constructor as typeof FormControl;
-      const validity: CustomValidityState = {};
-
+      const validity: Partial<Record<keyof ValidityState, boolean>> = {};
       let validationMessage = '';
       let isValid = true;
 
       proto.validators.forEach((validator) => {
-        /** Get data oof the Validator */
+        /** Get data off the Validator */
         const { message, callback } = validator;
 
         /** If a key is not set, use `customError` as a catch-all */
@@ -310,7 +447,7 @@ export function FormControlMixin<
           isValid = false;
           let messageResult = '';
 
-           /**
+          /**
            * The Validator interfaces allows for the message property
            * to be either a string or a function. If it is a function,
            * we want to get the returned value to use when calling
@@ -385,19 +522,21 @@ export function FormControlMixin<
       }
     }
 
-     /** Reset control state when the form is reset */
-    formResetCallback() {
-      this.setValue(null);
-      this.#touched = false;
-      this.#forceError = false;
-      this.#shouldShowError();
-      this?.resetFormControl();
-    }
+    /**
+     * This method is used to override the controls' validity message
+     * for a given Validator key. This has the highest level of priority when
+     * setting a validationMessage, so use this method wisely.
+     *
+     * The returned value will be used as the validationMessage for the given key.
+     * @param validationKey {string} - The key that has returned invalid
+     */
+    declare validityCallback: (validationKey: string) => string | void;
 
     /**
-     * Resets a form control to its initial state
+     * Called when the control's validationMessage should be changed
+     * @param message { string } - The new validation message
      */
-    declare resetFormControl: () => void;
+    declare validationMessageCallback: (message: string) => void;
   }
 
   return FormControl as Constructor<FormControlInterface> & TBase;
